Safari-611.1.10/LayoutTests/http/tests/security/xss-DENIED-invalid-domain-change-expected.txt

CONSOLE MESSAGE: SecurityError: Attempted to use a non-registrable domain.

PASS: cross-site not access allowed
